move main page to separate component
