+++ src/Landing/Explore/Card.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { CardLink, CardImg } from "../styled";
+import { CardLink, CardImg } from "../../styled";
 
 const Card = styled(CardLink)`
   border: 1px solid rgba(72, 72, 72, 0.2);
