+++ src/Landing/Homes/Card.js
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 
-import { CardLink } from "../styled";
-import Stars from "../Stars";
+import { CardLink } from "../../styled";
+import Stars from "../../Stars";
 
 const CardImg = styled.img`
   max-width: 100%;
