+++ src/Landing/Homes/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Col } from "react-flexbox-grid";
-import { TitleH2, HeaderRow, CarouselRow, CarouselArrow } from "../styled";
-import SeeAllLink from "../SeeAllLink";
+import { TitleH2, HeaderRow, CarouselRow, CarouselArrow } from "../../styled";
+import SeeAllLink from "../../SeeAllLink";
 
 import Card from "./Card";
 import image1 from "./1.png";
